Introduce DocsJsApiService and move GetInnerTextByIdAsync to it
